Add a comma between strings in format()
I tested the following code in GFortran:

    print 10
    print 11
    10 format('abc' 'de' 'f' 'g' 'h')
    11 format('abc','de','f','g','h')
    end

and it prints:

    abcdefgh
    abcdefgh

So it seems the 'a' 'b' is equivalent to 'a','b'. The latter compiles
with the latest NAG compiler and it seems to be standard Fortran 95,
unlike my previously proposed change in #16.

This is the only change needed to get SNAP compile with NAG, everything
else works.

Fixes #15.